section 4 + updates on section 2 and 5
finished the description of the component interfaces and sequence diagrams. Improved the integration and testing plan. Changes in the sequence diagrams of the data subscription that is now being handled in different ways depending on the type of the request, and the data upload/individual data request: now the DB stores the coordinates instead of the map names